Restructure so pygame runs in the __init__
__init__ - now has the main pygame loop
ArdunioCommands - Now has the classes for sending data to the ardunio to
be processed and passed on to devices. MotorSpeedSend class is included
here.
OSC - This is the main untouched library for OSC commands
OSC_sampleCode - is an proof of principle example that ultimately needs
to probably be split into its own module, with the results then being
brought into the __init__ to be processed after the pygame data.