OSC Server and PYGame now running together in __init__ - clashes of data
Both pygame while loop and OSC Server are now running in the same
programme successfully, but OSC Server is running in a separate thread
and as such can send data at the same time as the PYgame while loop.
This creates jumbled messages, which the Ardunio will not be able to
handle.

Maybe consider "locks" or "mutex" which hopefully will be able to lock
out one process, so that it can complete sending a message before the
other process/thread can start sending its message... hopefully.