Thread Locking Implemented to Stop Messages getting Muddled
Thread locking now implemented with "with lock:". This allows complete
OSC and complete pygame messages to be sent through the XBee without
them interfering or crossing over.

Next stage is to implement the OSC to do something simple like turn on
or off a LED, or move a simple servo with a fader.

2 Faders could also be used to set the tracking on the DC motors if they
look to be out of sync.... which inevitably they will.